Split range and lazy tag internals
