+++ include/cbor_tags/cbor.h
@@ -4,6 +4,7 @@
 
 // Float 16, c++23 has std::float16_t from <stdfloat> maybe, for now use float16_t below
 #include "cbor_tags/cbor_concepts.h"
+#include "cbor_tags/cbor_ranges.h"
 #include "cbor_tags/cbor_simple.h"
 #include "cbor_tags/float16_ieee754.h"
 
@@ -272,43 +273,6 @@ struct as_map {
     std::uint64_t size_;
     constexpr as_map(std::uint64_t size) : size_(size) {}
 };
-
-template <std::ranges::view R> struct array_range {
-    R range_;
-
-    constexpr explicit array_range(R range) : range_(std::move(range)) {}
-};
-
-template <std::ranges::view R> struct map_range {
-    R range_;
-
-    constexpr explicit map_range(R range) : range_(std::move(range)) {}
-};
-
-template <std::ranges::view R> struct bstr_range {
-    R           range_;
-    std::size_t chunk_size_{4096};
-
-    constexpr explicit bstr_range(R range, std::size_t chunk_size = 4096) : range_(std::move(range)), chunk_size_(chunk_size) {}
-};
-
-template <std::ranges::viewable_range R>
-    requires IsCborMajor<std::ranges::range_value_t<std::views::all_t<R>>>
-[[nodiscard]] constexpr auto as_array_range(R &&range) {
-    return array_range<std::views::all_t<R>>{std::views::all(std::forward<R>(range))};
-}
-
-template <std::ranges::viewable_range R>
-    requires IsPairLikeRange<std::views::all_t<R>>
-[[nodiscard]] constexpr auto as_map_range(R &&range) {
-    return map_range<std::views::all_t<R>>{std::views::all(std::forward<R>(range))};
-}
-
-template <std::ranges::viewable_range R>
-    requires IsByteLikeRange<std::views::all_t<R>>
-[[nodiscard]] constexpr auto as_bstr_range(R &&range, std::size_t chunk_size = 4096) {
-    return bstr_range<std::views::all_t<R>>{std::views::all(std::forward<R>(range)), chunk_size};
-}
 
 struct as_indefinite_map {};
 struct end_map {};
