+++ include/cbor_tags/cbor_detail.h
@@ -1,6 +1,8 @@
 #pragma once
 
 #include "cbor_tags/cbor_concepts.h"
+#include "cbor_tags/cbor_reflection_config.h"
+#include "cbor_tags/cbor_reflection_count.h"
 
 #include <concepts>
 #include <cstddef>
@@ -11,13 +13,6 @@
 #include <stdexcept>
 #include <tuple>
 #include <type_traits>
-
-#if defined(__cpp_impl_reflection) && __cpp_impl_reflection >= 202506L
-#include <meta>
-#define CBOR_TAGS_HAS_STD_REFLECTION 1
-#else
-#define CBOR_TAGS_HAS_STD_REFLECTION 0
-#endif
 
 namespace cbor::tags::detail {
 
@@ -189,20 +184,7 @@ template <typename Tuple> constexpr auto tuple_tail(Tuple &&tuple) {
                       std::forward<Tuple>(tuple));
 }
 
-#if CBOR_TAGS_HAS_STD_REFLECTION
-
-template <typename T>
-    requires IsAggregate<T> || IsTuple<T>
-constexpr auto aggregate_binding_count = [] consteval {
-    using type = std::remove_cvref_t<T>;
-    if constexpr (IsTuple<type>) {
-        return std::tuple_size_v<type>;
-    } else {
-        return std::meta::nonstatic_data_members_of(^^type, std::meta::access_context::current()).size();
-    }
-}();
-
-#else
+#if !CBOR_TAGS_HAS_STD_REFLECTION
 
 template <typename T, typename... TArgs>
     requires IsAggregate<T> || IsTuple<T>
@@ -231,7 +213,7 @@ template <typename Pair> constexpr decltype(auto) pair_first(Pair &&pair) {
     if constexpr (requires { std::get<0>(std::forward<Pair>(pair)); }) {
         return std::get<0>(std::forward<Pair>(pair));
     } else {
-        return std::forward<Pair>(pair).first;
+        return (std::forward<Pair>(pair).first);
     }
 }
 
@@ -239,7 +221,7 @@ template <typename Pair> constexpr decltype(auto) pair_second(Pair &&pair) {
     if constexpr (requires { std::get<1>(std::forward<Pair>(pair)); }) {
         return std::get<1>(std::forward<Pair>(pair));
     } else {
-        return std::forward<Pair>(pair).second;
+        return (std::forward<Pair>(pair).second);
     }
 }
