+++ include/cbor_tags/cbor_lazy_tags.h
@@ -1,0 +1,3 @@
+#pragma once
+
+#include "cbor_tags/cbor_lazy_tags_impl.h"
