+++ include/cbor_tags/cbor_reflection_config.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#if defined(__cpp_impl_reflection) && __cpp_impl_reflection >= 202506L
+#include <meta>
+#define CBOR_TAGS_HAS_STD_REFLECTION 1
+#else
+#define CBOR_TAGS_HAS_STD_REFLECTION 0
+#endif
