+++ test/test_header_split_reflection.cpp
@@ -1,0 +1,7 @@
+#include <cbor_tags/cbor_reflection_config.h>
+#include <cbor_tags/cbor_reflection_count.h>
+#include <cbor_tags/cbor_reflection_std.h>
+
+#include <doctest/doctest.h>
+
+TEST_CASE("reflection split headers are directly includable") { CHECK(true); }
